completed cicero for the most part
Generally, there are still a couple bugs left in there, but it's a lot less than there used to be. Considering that, this time, we have all the text and their locations generally in there, I would believe it's time to move on from Cicero for the time being.
The remaining bugs are a few minor issues like Roman numerals popping up in various places, but for the huge majority of the code scrapes text correctly.
And also looking at Cicero all the time is a terrible thing in general. Definitely time to move on.