Updated part of the README to include links
This is mostly a test to see if these links work and still make the chart appear legible. As a result, only part of the README.md now contains hyperlinks to respective collections, alongside a link to the code in the repository.
The rest of the table will follow before the week's end.