Caecilius Balbus Extractor
Katy was originally working on this one, but it was giving her problems
so I took over. And then it was giving me problems, since its
organizational method is really strange. However, I managed to pull
something together. Might need to be changed later to account for verse
number, but I believe it is acceptable in its current state.