handful of medieval changes
- Henry of Settimello: documentType changed to poetry; changed
collectiontitle and date
- Historia Brittonum: changed the code so that verses are properly
numbered according to the subsection number in the text. In the
database, multiple entries may have the same verse number because they
are in the same subsection. Perhaps, someday, we will find a way to
merge paragraphs before putting them in the database.
- Inquisitio: removed the date since it is unknown.
- Legenda Stephani: replaced parse function with case 1 from the
template since I believe it works better with the section numbers here.
- fixed an import in the template